+++ src/ham_fisted/function.clj
@@ -2,8 +2,10 @@
   "Helpers for working with [java.util.function](https://docs.oracle.com/javase/8/docs/api/java/util/function/package-summary.html)
   package objects."
   (:import [ham_fisted IFnDef$ODO IFnDef$OOO IFnDef$OO IFnDef$OL IFnDef$LongPredicate
-            IFnDef$DoublePredicate IFnDef$DD IFnDef$LL IFnDef IFnDef$LD IFnDef$DL IFnDef$OD]
-           [java.util.function BiFunction BiConsumer Function DoublePredicate LongPredicate Predicate]
+            IFnDef$DoublePredicate IFnDef$DD IFnDef$LL IFnDef IFnDef$LD IFnDef$DL IFnDef$OD
+            IFnDef$LO Casts IFnDef$Predicate]
+           [java.util.function BiFunction BiConsumer Function DoublePredicate LongPredicate Predicate
+            Consumer LongConsumer DoubleConsumer]
            [java.util Comparator]
            [clojure.lang Util]
            [it.unimi.dsi.fastutil.longs LongComparator]