+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-# 1.000-beta-94
+# 1.000-beta-96
+ * Fixed from upgrading dtype-next.
  * Major breaking changes!! API functions have been moved to make the documentation
    clearer and the library more maintainable in the long run.
  * map-union of mutable or transient maps produces mutable or transient maps!
@@ -6,7 +7,7 @@
  * Functions to make creating java.util.function objects are moved to ham-fisted.function.
  * Reduction-related systems are moved to ham-fisted.reduce.
  * java.util.Map helpers are moved to ham-fisted.mut-map.
-
+ 
 # 1.000-beta-93
  * java implementation of a batched stream reducer.  This avoids adding java to
    the stream api.